requirements.txt
d2to1>=0.2.10
pbr>=0.5.16

oslo.config>=1.1.0
simplejson
